requirements.txt
asgiref
autopep8
django
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
Pillow
django-bootstrap-form
django-import-export
django-jalali-date
django-extra-settings
django-ckeditor
graphene
graphene-django
django_jalali
unicodecsv
